requirements.txt
fastapi
uvicorn
pandas
matplotlib
beautifulsoup4
requests
scipy
lxml
google-generativeai
pandas
numpy
matplotlib
seaborn
scikit-learn
requests
PyMuPDF
networkx
duckdb
playwright-sync
bs4
openai
pillow
tqdm
pyarrow
pydantic
pypdf
